Move functions into banner.cpp
